move cmd/ctr-remote/{logger,sampler} to optimizer/{logger,sampler}
The optimizer packages should be decoupled from the CLI concept.
