+++ optimizer/sampler/options.go
